Move `ValidationErrors` to `validation`
