+++ crates/fj-core/src/validate/mod.rs
@@ -81,7 +81,7 @@ pub use self::{
     sketch::SketchValidationError, solid::SolidValidationError,
 };
 
-use std::{collections::HashMap, error::Error, fmt, thread};
+use std::{collections::HashMap, error::Error, thread};
 
 use fj_math::Scalar;
 
@@ -205,23 +205,5 @@ impl Default for ValidationConfig {
             // adjust it.
             identical_max_distance: Scalar::from_f64(5e-14),
         }
-    }
-}
-
-/// A collection of validation errors
-#[derive(Debug, thiserror::Error)]
-pub struct ValidationErrors(pub Vec<ValidationError>);
-
-impl fmt::Display for ValidationErrors {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let num_errors = self.0.len();
-
-        writeln!(f, "{num_errors} unhandled validation errors:")?;
-
-        for err in &self.0 {
-            writeln!(f, "{err}")?;
-        }
-
-        Ok(())
     }
 }
