+++ crates/fj-core/src/validation/error.rs
@@ -1,4 +1,4 @@
-use std::convert::Infallible;
+use std::{convert::Infallible, fmt};
 
 use crate::validate::{
     CycleValidationError, EdgeValidationError, FaceValidationError,
@@ -36,5 +36,23 @@ pub enum ValidationError {
 impl From<Infallible> for ValidationError {
     fn from(infallible: Infallible) -> Self {
         match infallible {}
+    }
+}
+
+/// A collection of validation errors
+#[derive(Debug, thiserror::Error)]
+pub struct ValidationErrors(pub Vec<ValidationError>);
+
+impl fmt::Display for ValidationErrors {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let num_errors = self.0.len();
+
+        writeln!(f, "{num_errors} unhandled validation errors:")?;
+
+        for err in &self.0 {
+            writeln!(f, "{err}")?;
+        }
+
+        Ok(())
     }
 }
