+++ crates/fj/src/instance.rs
@@ -6,7 +6,8 @@ use fj_core::{
         bounding_volume::BoundingVolume,
         triangulate::Triangulate,
     },
-    validate::{ValidationConfig, ValidationErrors},
+    validate::ValidationConfig,
+    validation::ValidationErrors,
     Core,
 };
 use fj_interop::Model;
